Implementation of mappers in remote layer
