+++ remote/src/main/java/com/universodoandroid/starwarsjetpack/remote/di/RemoteModules.kt
@@ -2,7 +2,7 @@ package com.universodoandroid.starwarsjetpack.remote.di
 
 import com.universodoandroid.starwarsjetpack.remote.BuildConfig
 import com.universodoandroid.starwarsjetpack.remote.ApiDataSource
-import com.universodoandroid.starwarsjetpack.remote.people.data.di.getPeopleRemoteModules
+import com.universodoandroid.starwarsjetpack.remote.people.di.getPeopleRemoteModules
 import org.koin.dsl.module
 
 internal val apiDataSource = module {
