+++ remote/src/main/java/com/universodoandroid/starwarsjetpack/remote/people/data/PeopleRemoteDataImpl.kt
@@ -1,13 +1,15 @@
-package com.universodoandroid.starwarsjetpack.remote.people.source
+package com.universodoandroid.starwarsjetpack.remote.people.data
 
 import com.universodoandroid.starwarsjetpack.data.people.datastore.PeopleRemoteData
 import com.universodoandroid.starwarsjetpack.data.people.entities.PeoplePageData
-import com.universodoandroid.starwarsjetpack.remote.people.mapper.PersonMapper
-import com.universodoandroid.starwarsjetpack.remote.people.data.PeopleRemoteDataSource
+import com.universodoandroid.starwarsjetpack.remote.people.remote.PeopleRemoteDataSource
+import com.universodoandroid.starwarsjetpack.remote.people.remote.response.PeopleResponse
+import com.universodoandroid.starwarsjetpack.shared.mapper.Mapper
 import io.reactivex.Flowable
 
 internal class PeopleRemoteDataImpl(
-    private val peopleRemoteDataSource: PeopleRemoteDataSource
+    private val peopleRemoteDataSource: PeopleRemoteDataSource,
+    private val mapper: Mapper<PeopleResponse, PeoplePageData>
 ) : PeopleRemoteData {
 
     override fun getAllPeopleData(): Flowable<PeoplePageData> = sync(1)
@@ -23,9 +25,6 @@ internal class PeopleRemoteDataImpl(
     }
 
     override fun getPeoplePerPage(page: Int): Flowable<PeoplePageData> {
-        return peopleRemoteDataSource.loadPeoplePerPage(page).map {
-            PersonMapper.responseToPeoplePage(it)
-        }
+        return peopleRemoteDataSource.loadPeoplePerPage(page).map { mapper.map(it) }
     }
-
 }
