--- remote/src/main/java/com/universodoandroid/starwarsjetpack/remote/people/mapper/PersonMapper.kt
@@ -1,40 +1,0 @@
-package com.universodoandroid.starwarsjetpack.remote.people.mapper
-
-import com.universodoandroid.starwarsjetpack.data.people.entities.PeoplePageData
-import com.universodoandroid.starwarsjetpack.data.people.entities.PersonData
-import com.universodoandroid.starwarsjetpack.remote.people.data.response.PeopleResponse
-import com.universodoandroid.starwarsjetpack.remote.people.data.response.PersonItem
-
-object PersonMapper {
-
-    private fun fromResponse(personItem: PersonItem) =
-        PersonData(
-            id = "",
-            birthYear = personItem.birthYear,
-            created = personItem.created,
-            edited = personItem.edited,
-            eyeColor = personItem.eyeColor,
-            gender = personItem.gender,
-            hairColor = personItem.hairColor,
-            height = personItem.height,
-            homeworld = personItem.homeworld,
-            mass = personItem.mass,
-            name = personItem.name,
-            skinColor = personItem.skinColor,
-            url = personItem.url
-        )
-
-    fun responseToPeoplePage(peopleResponse: PeopleResponse): PeoplePageData {
-        val hasNext = !peopleResponse.next.isNullOrBlank()
-
-        return PeoplePageData(
-            hasNextPage = hasNext,
-            people = responseToEntities(
-                peopleResponse.results
-            )
-        )
-    }
-
-    private fun responseToEntities(response: List<PersonItem>) = response.map(PersonMapper::fromResponse)
-
-}
