+++ remote/src/test/java/com/universodoandroid/starwarsjetpack/remote/observers/FakePeopleRemoteDataSource.kt
@@ -2,7 +2,7 @@ package com.universodoandroid.starwarsjetpack.remote.observers
 
 import com.universodoandroid.starwarsjetpack.remote.people.api.PeopleApiDataSource
 import com.universodoandroid.starwarsjetpack.remote.data.PeopleReponseDataFactory
-import com.universodoandroid.starwarsjetpack.remote.people.data.response.PeopleResponse
+import com.universodoandroid.starwarsjetpack.remote.people.remote.response.PeopleResponse
 import io.reactivex.BackpressureStrategy
 import io.reactivex.Flowable
 import io.reactivex.Observable
